# ml_api/app/swagger_template.yml
openapi: 3.0.3
info:
  title: Match Vagas x Candidatos API
  version: 1.0.0
  description: >
    API REST para cadastrar candidatos, cadastrar vagas e gerar
    ranking dos 10 melhores candidatos para uma vaga específica.
servers:
  - url: http://localhost:5000
paths:
...
  /match/{vaga_id}:
    get:
      summary: Rankeia top10 candidatos para vaga
      tags:
        - Match
      parameters:
        - in: path
          name: vaga_id
          required: true
          schema:
            type: integer
            example: 42
          description: ID da vaga
      responses:
        '200':
          description: Lista dos 10 melhores candidatos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchEntry'
        '404':
          description: Vaga não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
...
    MatchEntry:
      type: object
      properties:
        id:
          type: integer
          example: 123
        nome:
          type: string
          example: João Pereira
        email:
          type: string
          format: email
          example: joao.pereira@exemplo.com
        compatibilidade:
          type: number
          format: float
          example: 87.65
    Error:
      type: object
      properties:
        error:
          type: string
          example: Mensagem de erro explicativa